requirements.txt
crewai
langchain
langchain_community
langchain_core
openai
panel
python-dotenv
